fix(crates/arkworks): the pairing tests embedded in curves
